WIP commands reshuffle
