+++ brewblox_ctl_lib/loader.py
@@ -2,10 +2,14 @@
 Collects and returns all CLI groups in the correct order
 """
 
-from brewblox_ctl_lib import commands
+from brewblox_ctl_lib.commands import backup, debug, service, setup, update
 
 
 def cli_sources():
     return [
-        commands.cli,
+        setup.cli,
+        service.cli,
+        update.cli,
+        debug.cli,
+        backup.cli,
     ]
